webserver/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
aiofiles>=23.2.1
websockets>=12.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
jinja2>=3.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
watchdog>=3.0.0